Refactor entities and events; add domain and aggregate interfaces
Refactored `TestEntity` to use constructor initialization for `Id` and changed it from a record to a class. Updated `EventEntity` and `Entity` to be generic with `TKey` and `TTimeStamp` type parameters. Changed `EventVersion` type to `ulong` in `Event`, `IEvent`, and `IEventEntity`. Added new properties to `IEntity` and removed `IEntityConcurrency` and `IEntityRelational` interfaces. Introduced `EventDomain`, `IEventDomain`, `IAggregate`, and `IEventConverter` interfaces for domain events and aggregates.

+++ Xpandables.Net/Events/Event.cs
@@ -25,7 +25,7 @@ public abstract record Event : IEvent
     public DateTimeOffset OccurredOn { get; init; } = DateTimeOffset.UtcNow;
 
     /// <inheritdoc/>
-    public int EventVersion { get; init; }
+    public ulong EventVersion { get; init; }
 
     /// <inheritdoc/>
     public Guid EventId { get; init; } = Guid.NewGuid();
